Replace all umlauts with character entities.
This is not only desirable when using this in conjuction with pdflatex,
but also makes it easier for people using alternative keyboard layouts
that do not have umlaut keys.

While replacing the umlauts manually (I thought it wasn't worth creating
a big search+replace), I found and fixed two errors:

 * The reference to "Steuerbeguenstigte Zwecke" was quoted incorrectly
   and also the case in the original ODT document.
 * A % wasn't properly escaped (% are TeX comments and so the remainder
   of the line was omitted).

I've rewrapped the lines to 80 characters, because the replacement for
single character umlauts with entities consisting of two characters also
increases the line length.

Closes #2.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>